Allow `$` & `~` inside paths in galaxy roles (#72966)
ansible-galaxy currently behaves bad then a role to be installed
contains `~` or `$` at any place in the path of a file in that role.
It extracts the parent directory of the offending path part as an
empty file. This explodes if that directory contains anything else.

Changed this behaviour. Those symbols are now allowed in the paths
because I have no idea why they forbidden in the first case.
Moved path sanitizing into a dedicated function that checks for
suspicious paths. If a path looks odd, the installation fails
loudly instead of just extracting nothing into half the path.
This could break compatibility but I do not see which use case
this change would prevent. The check can be bypassed by passing
`--force` as argument to `ansible-galaxy`.

While fixing that another bug showed up: ansible-galaxy tries
to install roles to different directories. If all of them are
non writable, an unexcepted Exception occurs. Code had a lot of
spaghetti in it. Moved extraction to extra method and actuall
 clean up if an installation fails.

Co-authored-by: Sviatoslav Sydorenko <wk.cvs.github@sydorenko.org.ua>